jwt login

+++ backend/controllers/contactControllers.js
@@ -4,28 +4,28 @@ const asyncHandler=require("express-async-handler");
 // @route GET /api/contacts
 // @acess public
 const getsContact=asyncHandler(async(req,res)=>{
-    const contacts= await Contact.find();
+    const contacts= await Contact.find({user_id:req.user.id});
 
     res.status(200).json(contacts);
 })
 // @desc create a contacts
 // @route POST /api/contacts/:id
 // @acess public
-const createContact=asyncHandler(async(req,res)=>{
-    const {name,number}=req.body;
-     if(!name|| !number){
+const createContact = asyncHandler(async (req, res) => {
+    const { name, number } = req.body;
+    if (!name || !number) {
         res.status(400);
         throw new Error("all fields are mandatory");
     }
-    const contact=await Contact.create({
+    const contact = await Contact.create({
         name,
         number,
+        user_id: req.user.id
     });
-    res.status(200).json(contact );
-    console.log("the body is",req.body);
-    
-   
-})
+    console.log("Contact created:", contact);
+    res.status(200).json(contact);
+});
+
 // @desc update a contact
 // @route PUT /api/contacts/:id
 // @acess public
@@ -34,6 +34,10 @@ const updateContact=asyncHandler(async (req,res)=>{
     if(!contact){
         res.status(404);
         throw new Error("contact not found");
+    }
+    if(contact.user_id.toString()!== req.user.id){
+        res.status(403);
+        throw new Error("you cant change this contacts")
     }
     const updateContact=await Contact.findByIdAndUpdate(
         req.params.id,
@@ -50,6 +54,10 @@ const deleteContact= asyncHandler(async(req,res)=>{
     if(!contact){
         res.status(404);
         throw new Error("contact not found");
+    }
+     if(contact.user_id.toString()!== req.user.id){
+        res.status(403);
+        throw new Error("you cant change this contacts")
     }
     console.log("Fetched contact:", contact); 
     await contact.deleteOne();
